+++ km3db/cli/detx.py
@@ -4,15 +4,16 @@ Retrieves DETX files from the database.
 
 Usage:
     detx [options] DET_ID
-    detx DET_ID RUN
+    detx [options] DET_ID RUN
     detx (-h | --help)
     detx --version
 
 Options:
     DET_ID        The detector ID (e.g. 49)
     RUN           The run ID.
-    -c CALIBR_ID  Geometrical calibration ID (eg. A01466417)
-    -t T0_SET     Time calibration ID (eg. A01466431)
+    -p P_CAL      Position calibration ID [default: 0].
+    -r R_CAL      Rotation calibration ID [default: 0].
+    -t T_CAL      Time calibration ID [default: 0].
     -o OUT        Output folder or filename.
     -h --help     Show this screen.
 
@@ -38,7 +39,9 @@ def main():
     if args["RUN"] is not None:
         detx = km3db.tools.detx_for_run(det_id, int(args["RUN"]))
     else:
-        detx = km3db.tools.detx(det_id, t0set=args["-t"], calibration=args["-c"])
+        detx = km3db.tools.detx(
+            det_id, pcal=args["-p"], rcal=args["-r"], tcal=args["-t"]
+        )
 
     if detx is None:
         log.error("No detx found.")